Added Auth-service, restructured to have microservices
Moved all services down, deleted the overarching backend go module, and
brought the db/models dependencies over into main-service.

Auth service now launches with docker-compose up
