+++ backend/services/auth-service/cmd/main.go
@@ -1,9 +1,9 @@
 package main
 
 import (
-	"backend/auth-service/api"
-	"backend/auth-service/internal/middleware"
 	"database/sql"
+	"github.com/jordyob03/TripTailor/backend/services/auth-service/api"
+	"github.com/jordyob03/TripTailor/backend/services/auth-service/internal/middleware"
 	"log"
 	"net/http"
 	"time"
