+++ backend/services/auth-service/internal/models/user.go
