+++ backend/services/itinerary-service/internal/models/event.go
