+++ backend/services/itinerary-service/internal/models/itinerary.go
