+++ backend/services/main-service/go.mod
@@ -1,0 +1,5 @@
+module github.com/jordyob03/TripTailor/backend/services/main-service
+
+go 1.23
+
+require github.com/lib/pq v1.10.9
